requirements.txt
faster-whisper
piper-tts
langchain-google-genai
pyaudio
numpy
webrtcvad
onnxruntime
